+++ fortscale/presidio-input/presidio-input-core/src/main/java/presidio/input/core/services/impl/InputExecutionServiceImpl.java
@@ -30,7 +30,7 @@ public class InputExecutionServiceImpl implements PresidioExecutionService {
     }
 
     @Override
-    public void run(DataSource dataSource, Instant startDate, Instant endDate, Long fixedDuration) throws Exception {
+    public void run(DataSource dataSource, Instant startDate, Instant endDate, Double fixedDuration) throws Exception {
         logger.info("Started input processing with params: data source:{}, from {}:{}, until {}:{}.",dataSource, CommonStrings.COMMAND_LINE_START_DATE_FIELD_NAME, startDate, CommonStrings.COMMAND_LINE_END_DATE_FIELD_NAME, endDate);
 
         final List<? extends AbstractAuditableDocument> dataRecords = find(dataSource, startDate, endDate);
